requirements.txt
fastapi
uvicorn
openai
loguru
python-dotenv
httpx
python-multipart
streamlit
pandas
pydub
